mapped the database classes in java

+++ src/main/java/it/uiip/digitalgarage/roboadvice/persistence/repository/UserRepository.java
@@ -1,16 +1,12 @@
 package it.uiip.digitalgarage.roboadvice.persistence.repository;
 
-import it.uiip.digitalgarage.roboadvice.persistence.model.AssetClass;
 import it.uiip.digitalgarage.roboadvice.persistence.model.User;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
-import org.springframework.data.repository.query.Param;
 
 /**
  * Created by feder on 28/02/2017.
  */
-public interface UserRepository extends PagingAndSortingRepository<User, Long> {
+public interface UserRepository extends PagingAndSortingRepository<User, Integer> {
 
     User findByEmail(String email);
 }
